backend/requirements.txt
Flask
face_recognition
dlib
numpy
pillow